fix: repair compile error with XCode.

+++ Morty/Core/Math/MMath.cpp
@@ -1,7 +1,7 @@
 #include "Math/MMath.h"
 #include "Utility/MTimer.h"
 
-std::default_random_engine MMath::s_randomEngine(MTimer::GetCurTime());
+std::default_random_engine MMath::s_randomEngine(static_cast<uint32_t>(MTimer::GetCurTime()));
 
 Matrix4 MMath::GetScaleAndRotation(const Matrix4& mat)
 {
