+++ Morty/Editor/Property/PropertyBase.h
@@ -49,7 +49,7 @@ public:
 	bool EditVector4(Vector4& value, const float& fSpeed = 1.0f, const float& fMin = 0.0f, const float& fMax = 0.0f);
 	bool EditVector4(float* pValue, const float& fSpeed = 1.0f, const float& fMin = 0.0f, const float& fMax = 0.0f);
 	bool EditMTransform(MTransform& trans);
-	bool EditEnum(const std::vector<MString>& select, uint32_t& index);
+	bool EditEnum(const std::vector<MString>& select, size_t& index);
 	bool EditMColor(MColor& value);
 	bool EditMString(MString& value);
