+++ Morty/Editor/Widget/TaskGraphView.cpp
@@ -85,7 +85,7 @@ void TaskGraphView::Render()
 			vAllNodes.push_back(pNode);
 		}
 
-		for (int nInputIdx = pNode->GetInputSize() - 1; nInputIdx >= 0; --nInputIdx)
+		for (int nInputIdx = static_cast<int>(pNode->GetInputSize()) - 1; nInputIdx >= 0; --nInputIdx)
 		{
 			MTaskNode* pPrevNode = pNode->GetInput(nInputIdx)->GetLinkedNode();
 		    vNodes.push_back(pPrevNode);
@@ -94,7 +94,7 @@ void TaskGraphView::Render()
 
 	for (auto& pNode : vAllNodes)
 	{
-		const int imNodeId = reinterpret_cast<std::intptr_t>(pNode);
+		const int imNodeId = static_cast<int>(reinterpret_cast<std::intptr_t>(pNode));
 		(ImNodes::BeginNode(imNodeId));
 		{
 			//first initialize position.
@@ -118,7 +118,7 @@ void TaskGraphView::Render()
 			for (size_t nIdx = 0; nIdx < pNode->GetInputSize(); ++nIdx)
 			{
 				auto pNodeInput = pNode->GetInput(nIdx);
-				ImNodes::BeginInputAttribute( reinterpret_cast<std::intptr_t>(pNodeInput));
+				ImNodes::BeginInputAttribute( static_cast<int>(reinterpret_cast<std::intptr_t>(pNodeInput)));
 				ImGui::Text("%s", pNodeInput->GetName().c_str());
 				ImNodes::EndInputAttribute();
 			}
@@ -127,7 +127,7 @@ void TaskGraphView::Render()
 			for (size_t nIdx = 0; nIdx < pNode->GetOutputSize(); ++nIdx)
 			{
 				auto pNodeOutput = pNode->GetOutput(nIdx);
-				ImNodes::BeginOutputAttribute(reinterpret_cast<std::intptr_t>(pNodeOutput));
+				ImNodes::BeginOutputAttribute(static_cast<int>(reinterpret_cast<std::intptr_t>(pNodeOutput)));
 				ImGui::Text("%s", pNodeOutput->GetName().c_str());
 				ImNodes::EndOutputAttribute();
 			}
@@ -147,8 +147,8 @@ void TaskGraphView::Render()
 				continue;
 			}
 
-			const int inputId = reinterpret_cast<std::intptr_t>(pInput);
-			const int outputId = reinterpret_cast<std::intptr_t>(pOutput);
+			const int inputId = static_cast<int>(reinterpret_cast<std::intptr_t>(pInput));
+			const int outputId = static_cast<int>(reinterpret_cast<std::intptr_t>(pOutput));
 
 			ImNodes::Link(inputId + outputId, outputId, inputId);
 		}
