+++ Morty/Render/Culling/MCascadedShadowCulling.cpp
@@ -120,9 +120,9 @@ void MCascadedShadowCulling::Culling(const std::vector<MMaterialBatchGroup*>& vI
 				{
 					const MMeshManager::MMeshData& data = pMeshManager->FindMesh(instance.pMesh);
 					const MDrawIndexedIndirectData indirectData = {
-					data.indexInfo.size,
+                        static_cast<uint32_t>(data.indexInfo.size),
 							1,
-							data.indexInfo.begin,
+                        static_cast<uint32_t>(data.indexInfo.begin),
 							0,
 							static_cast<uint32_t>(nIdx)
 					};
